(TAB-9) Prepare for Braille updates by moving Braille validation into its own class and source file. (Otherwise unchanged.)
